requirements.txt
sounddevice
soundfile
numpy
deepgram-sdk
python-dotenv
google-generativeai
fastapi
uvicorn[standard]
python-multipart
